Separated the trackball code from keymap
