+++ keyboards/handwired/dactyl_manuform/5x6_r_track/5x6_r_track.h
@@ -25,6 +25,6 @@
     { R10, R11, R12, R13, R14, R15 }, \
     { R20, R21, R22, R23, R24, R25 }, \
     { R30, R31, R32, R33, R34, R35 }, \
-    { R40, R41, R42, R43, XXX, XXX }, \
-    { R50, R51, R52, R53, XXX, XXX } \
+    { XXX, R41, R42, R43, XXX, XXX }, \
+    { XXX, R51, R52, R53, XXX, XXX } \
 }
